The user products page was created. A separate component for the product list has been created. The products page has been renamed to official products.
