+++ backend/core/views/product.py
@@ -11,10 +11,7 @@ class ProductFilter(filters.FilterSet):
     # Talvez filtro pelos internal_names de release e product_type
     class Meta:
         model = Product
-        fields = [
-            "release",
-            "product_type",
-        ]
+        fields = ["release", "product_type", "official_product"]
 
 
 class ProductViewSet(viewsets.ModelViewSet):
